+++ tests/sql_test_files/it_works_protocol.sql
@@ -1,0 +1,6 @@
+select 'text' as component, 
+    CASE sqlpage.protocol()
+        WHEN 'http' THEN 'It works !'
+        ELSE 'It failed ! Expected "http", got "' || sqlpage.protocol() || '"".'
+    END
+    AS contents;